ファイル名更新、 mainAxisAlignment: MainAxisAlignment.start,消去、インデント修正
